Rejiggered a bunch of stuff for Android Studio
Idea is to use this directory for the Android Studio project and "workspace" for the Bazel workspace. You can import into Android Studio with workspace/.bazelproject and this repo will not contain any Android Studio project files but they should generate consistenly on import. Added a README.md that needs to be filled in more. Cleaned up a lot of the source code (old debug stuff, logs, visibility, etc.) and added myself to copyright notices. Fixed sharing. Tried out using lambdas but that does not seem to be working with Bazel so gave up on that. Tried Proguard one last time but that is also not working so gave up on that. Switched over to referencing support library things directly since that seems to make Android Studio pull in the "external libraries" a little more consistently.